Vis fritekstfelt mål for utdanning hvis heltid
+ litt refaktorering siden fritekstfeltet skl vises for både deltid og heltid. Forskjellen er da at deltid også skal ha et inputfelt for arbeidsmengde
